final project structure + assistant endpoint
