+++ src/api/models/get_request_models/assistant.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Assistant(BaseModel):
+    assistant_id: int
+    source:str
+    owner:str
