+++ src/api/routers/process_router.py
@@ -3,18 +3,17 @@ import aio_pika
 from aio_pika.abc import AbstractExchange, \
     DeliveryMode
 
-from domain.enteties.IOdataenteties.queue_enteties import FileTranscribationQuery
 from fastapi import APIRouter
 from starlette.requests import Request
 
-from src.api.models.post_request_models.start_process import StartFromYouTubeMessage
+from src.api.models.post_request_models.start_process import StartFromYouTubeMessage, StartFromStorageMessage
 
 processes_router = APIRouter()
 
 
 @processes_router.post("/api/start/start_process_from_storage")
-async def start_task(message: FileTranscribationQuery,
-                     request: Request):
+async def start_task_from_storage(message: StartFromStorageMessage,
+                                  request: Request):
     processor_exchange_object: AbstractExchange = request.app.state.process_exchange
     await processor_exchange_object.publish(
         aio_pika.Message(body=message.json().encode(),
@@ -23,8 +22,8 @@ async def start_task(message: FileTranscribationQuery,
 
 
 @processes_router.post("/api/start/start_process_from_youtube")
-async def start_task(message: StartFromYouTubeMessage,
-                     request: Request):
+async def start_task_from_youtube(message: StartFromYouTubeMessage,
+                                  request: Request):
     processor_exchange_object: AbstractExchange = request.app.state.process_exchange
     await processor_exchange_object.publish(
         aio_pika.Message(body=message.json().encode(),
