+++ src/api/routers/results_routers.py
@@ -3,12 +3,14 @@ from dataclasses import asdict
 from fastapi import HTTPException
 from fastapi import APIRouter
 from starlette.requests import Request
-from src.consumption.models.buisnes_models import TranscribedText, SummaryText
 
-database_router = APIRouter()
+from src.consumption.models.consumption.summarization import SummaryText
+from src.consumption.models.consumption.transcribition import TranscribedText
+
+results_router = APIRouter()
 
 
-@database_router.get("/api/results/get_transcribed_text")
+@results_router.get("/api/results/get_transcribed_text")
 async def get_transcribed_text_from_database(id_text: int,
                                              request: Request):
     try:
@@ -21,7 +23,7 @@ async def get_transcribed_text_from_database(id_text: int,
         raise HTTPException(status_code=500, detail=f"An error occurred: {str(e)}")
 
 
-@database_router.get("/api/results/get_summary_text")
+@results_router.get("/api/results/get_summary_text")
 async def get_summary_text_from_database(id_text: int,
                                          request: Request):
     try:
