+++ src/database/models/consumption/asssistant.py
@@ -1,5 +1,4 @@
 from sqlalchemy import Column, Integer, String, ForeignKey, TIMESTAMP, BigInteger, CHAR, Text, DECIMAL, TIME
-from sqlalchemy.dialects.postgresql import INTERVAL
 
 from src.database.models.base_model import ModelBase
