+++ src/database/models/consumption/summarization.py
@@ -1,7 +1,9 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, TIMESTAMP, BigInteger, CHAR, Text, DECIMAL, TIME
-from sqlalchemy.dialects.postgresql import INTERVAL
+from sqlalchemy import Column, Integer,  ForeignKey, TIMESTAMP, BigInteger,  Text,  TIME
+
+from sqlalchemy.orm import relationship
 
 from src.database.models.base_model import ModelBase
+
 
 class SummaryTexts(ModelBase):
     __tablename__ = 'summary_texts'
